chain, graph: Move EthereumAdapterError to graph::blockchain
