+++ graph/src/components/ethereum/mod.rs
@@ -6,10 +6,10 @@ mod types;
 
 pub use self::adapter::{
     blocks_with_triggers, triggers_in_block, BlockStreamMetrics, EthGetLogsFilter, EthereumAdapter,
-    EthereumAdapterError, EthereumBlockFilter, EthereumCallFilter, EthereumContractCall,
-    EthereumContractCallError, EthereumContractState, EthereumContractStateError,
-    EthereumContractStateRequest, EthereumLogFilter, EthereumNetworkIdentifier,
-    MockEthereumAdapter, ProviderEthRpcMetrics, SubgraphEthRpcMetrics,
+    EthereumBlockFilter, EthereumCallFilter, EthereumContractCall, EthereumContractCallError,
+    EthereumContractState, EthereumContractStateError, EthereumContractStateRequest,
+    EthereumLogFilter, EthereumNetworkIdentifier, MockEthereumAdapter, ProviderEthRpcMetrics,
+    SubgraphEthRpcMetrics,
 };
 pub use self::listener::{ChainHeadUpdate, ChainHeadUpdateListener, ChainHeadUpdateStream};
 pub use self::network::{EthereumNetworkAdapters, EthereumNetworks, NodeCapabilities};
